Syncing across this change from libwnck. Patch from Neil Muller
2004-04-19  Mark McLoughlin  <mark@skynet.ie>

        Syncing across this change from libwnck.
        Patch from Neil Muller <neil@dip.sun.ac.za> in bug #133979.

        * src/iconcache.c: (find_largest_sizes), (find_best_size):
        Don't down-size nitems from a gulong to an int. Fixes a
        crash with enlightenment, apparently.

+++ src/iconcache.c
@@ -46,7 +46,7 @@ get_fallback_icons (MetaScreen     *screen,
 
 static gboolean
 find_largest_sizes (gulong *data,
-                    int     nitems,
+                    gulong  nitems,
                     int    *width,
                     int    *height)
 {
@@ -81,7 +81,7 @@ find_largest_sizes (gulong *data,
 
 static gboolean
 find_best_size (gulong  *data,
-                int      nitems,
+                gulong   nitems,
                 int      ideal_width,
                 int      ideal_height,
                 int     *width,
